requirements.txt
neo4j>=5.20,<6
openai==1.0.0
pinecone>=3.0.0,<4
pyvis==0.3.1
networkx==3.1
tqdm
python-dotenv
